backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
smtplib
email-validator